Make BlendLuxCore reloadable in Blender.
The aim of this commit is to make BLC reload smoothly when a "Reload Scripts"
command is issued in Blender.
This ability is needed to have BLC in "editable" mode in Blender.

This is a big commit, but no other way to do was found, as the following tasks
had to be done:
- resolve circular references in imports. In particular, redispatch a few
  sub-submodules in the submodules
- correct some faulty syntactic constructions, like imports outside top level
  (see PEP 8, https://docs.astral.sh/ruff/rules/import-outside-top-level/,
  etc.)
- add the importlib.reload mechanics in modules (NB: to be continued, there
  could be missing ones)
- for some files, reformat according PEP8, to make refactoring easier
- for a very few submodules, fix some errors in the existing code that the
  linter detected when I made a final check
- add documentation.